Move TestTupleType into its own file
